packages/fantasy_mlb_ai/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fantasy-mlb-ai"
version = "0.1.0"
description = "Fantasy baseball management with ML-powered projections"
readme = "INTEGRATION_README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Austen Haymond"}
]

dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "espn-api>=0.31.0",
    "python-dotenv>=1.0.0",
    "matchup-machine",  # Workspace dependency
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
]

[tool.setuptools]
packages = ["fantasy_mlb_ai"]
package-dir = {"" = "src"}

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"
